added multiple document upload api

+++ app/services/user_service.py
@@ -1,9 +1,12 @@
 import uuid
+from fastapi.security import OAuth2PasswordRequestForm
 from sqlmodel.ext.asyncio.session import AsyncSession
+from app.core.security import verify_password
+from app.exceptions.security_exception import UnauthorizedUserException
 from app.exceptions.user_exceptions import UserNotFoundException
 from app.models.user import User
 from app.repository.user_repo import UserRepository
-from app.schema.user import UserLogin
+from app.schema.user import UserLogin, UserResponse
 
 
 class UserService:
@@ -14,8 +17,8 @@ class UserService:
 
 
     async def get_user_by_id(self, user_id: uuid.UUID):
-        result = await self.repo.get_user_by_id(self.session, user_id)
-        return result
+        user = await self.repo.get_user_by_id(self.session, user_id)
+        return UserResponse.model_validate(user)
 
 
 
@@ -28,15 +31,21 @@ class UserService:
         return await self.repo.get_all_users(self.session)
 
 
-    async def delete_user(self, user: User):
-        existing_user = await self.repo.get_user_by_id(self.session, user.id)
+
+    async def delete_user(self, user_data: OAuth2PasswordRequestForm):
+
+        existing_user = await self.repo.get_user_by_name(self.session, user_data.username)
         if existing_user is None:
-            raise UserNotFoundException(user.name)
+            raise UserNotFoundException(user_data.username)
+        
+        if not verify_password(user_data.password, existing_user.password):
+            raise UnauthorizedUserException(f"Unauthorized user : {user_data.username}")
         
         try:
             await self.repo.delete_user(self.session, existing_user)
             await self.session.commit()
+
         except Exception as e:
-            return str(e)
+            raise RuntimeError(f"Exception occured : {str(e)}") from e
